requirements.txt
# eCourts High Court Case Status Scraper Requirements

# Core web scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Async support
asyncio-throttle>=1.0.2

# Database

psycopg[binary]>=3.1.9  # PostgreSQL driver

# Environment management
python-dotenv>=1.0.0

# CAPTCHA solving
pytesseract>=0.3.10     # OCR-based CAPTCHA solving
Pillow>=9.0.0           # Image processing for OCR
numpy>=1.21.0           # Numerical operations for image preprocessing

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging and utilities
colorlog>=6.7.0

# Data processing
pandas>=2.0.0  # Optional, for data analysis
openpyxl>=3.1.0  # Optional, for Excel export
